requirements.txt
django==4.2.24
psycopg2-binary==2.9.9
python-telegram-bot==21.6